+++ bgw-gui/src/commonMain/kotlin/data/event/DragGestureEndedEventData.kt
@@ -20,4 +20,9 @@ package data.event
 import ID
 import kotlinx.serialization.Serializable
 
-@Serializable internal class DragGestureEndedEventData(var droppedOn: ID?) : EventData()
+@Serializable
+internal class DragGestureEndedEventData(
+    var droppedOn: ID?,
+    val posX: Double = 0.0,
+    val posY: Double = 0.0,
+) : EventData()