+++ bgw-gui/src/commonMain/kotlin/data/event/DragGestureStartedEventData.kt
@@ -19,4 +19,8 @@ package data.event
 
 import kotlinx.serialization.Serializable
 
-@Serializable internal class DragGestureStartedEventData : EventData()
+@Serializable
+internal class DragGestureStartedEventData(
+    val posX: Double = 0.0,
+    val posY: Double = 0.0,
+) : EventData()